Moved out functions used to interact with spotify into spotify_api.js
